Move editor/ into views/
